fix: 나머지 bigint를 tinyint로 수정했습니다
round, lotto_combination 테이블의 6개 번호 컬럼을 Byte 형식으로 바꾸고 서비스쪽도 에러가 없도록 수정하였습니다
로컬DB에서 정상으로 작동하는것을 확인하였고 DB에서 아래 명령어를 입력해서 테이블을 수정해주세요
ALTER TABLE round MODIFY COLUMN bonus tinyint NOT NULL;
ALTER TABLE round MODIFY COLUMN num1 tinyint NOT NULL;
ALTER TABLE round MODIFY COLUMN num2 tinyint NOT NULL;
ALTER TABLE round MODIFY COLUMN num3 tinyint NOT NULL;
ALTER TABLE round MODIFY COLUMN num4 tinyint NOT NULL;
ALTER TABLE round MODIFY COLUMN num5 tinyint NOT NULL;
ALTER TABLE round MODIFY COLUMN num6 tinyint NOT NULL;

ALTER TABLE lotto_combination MODIFY COLUMN first_num tinyint NOT NULL;
ALTER TABLE lotto_combination MODIFY COLUMN second_num tinyint NOT NULL;
ALTER TABLE lotto_combination MODIFY COLUMN third_num tinyint NOT NULL;
ALTER TABLE lotto_combination MODIFY COLUMN fourth_num tinyint NOT NULL;
ALTER TABLE lotto_combination MODIFY COLUMN fifth_num tinyint NOT NULL;
ALTER TABLE lotto_combination MODIFY COLUMN sixth_num tinyint NOT NULL;

issue: #32

+++ src/main/java/com/lotto/lotto_simulator/entity/LottoCombination.java
@@ -21,21 +21,21 @@ public class LottoCombination {
     private Long lotto_id;
 
     @Column(nullable = false)
-    private Long firstNum;
+    private Byte firstNum;
 
     @Column(nullable = false)
-    private Long secondNum;
+    private Byte secondNum;
 
     @Column(nullable = false)
-    private Long thirdNum;
+    private Byte thirdNum;
 
     @Column(nullable = false)
-    private Long fourthNum;
+    private Byte fourthNum;
 
     @Column(nullable = false)
-    private Long fifthNum;
+    private Byte fifthNum;
 
     @Column(nullable = false)
-    private Long sixthNum;
+    private Byte sixthNum;
 
 }
